+++ src/main/scala/org/frankversnel/poortjes/util/delta_time.scala
@@ -1,6 +1,6 @@
 package org.frankversnel.poortjes.util
 
-case class DeltaTime(millis: Int) {
+case class DeltaTime(millis: Long) {
 	val seconds: Float = millis.toFloat / 1000f
 }
